% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{cdc}
\alias{cdc}
\title{Fetch CDC page hits}
\usage{
cdc(auth, epiweeks, locations)
}
\arguments{
\item{auth}{auth token}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{dephi_states}}
}
\description{
Fetch CDC page hits
}
