% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{delphi}
\alias{delphi}
\title{Fetch Delphi's forecast}
\usage{
delphi(system, epiweek)
}
\arguments{
\item{system}{(not sure)}

\item{epiweek}{an epiweek (fomat \code{YYYYWW})}
}
\description{
Fetch Delphi's forecast
}
