% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi-articles.R
\docType{data}
\name{delphi_articles}
\alias{delphi_articles}
\title{Supported Wikipedia Article Titles}
\format{An object of class \code{character} of length 54.}
\usage{
delphi_articles
}
\description{
Supported Wikipedia Article Titles
}
\keyword{datasets}
