% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi-regions.R
\docType{data}
\name{delphi_regions}
\alias{delphi_regions}
\title{Supported Regions}
\format{An object of class \code{character} of length 20.}
\usage{
delphi_regions
}
\description{
Supported Regions
}
\keyword{datasets}
