% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi-states.R
\docType{data}
\name{delphi_states}
\alias{delphi_states}
\title{Supported States}
\format{An object of class \code{character} of length 51.}
\usage{
delphi_states
}
\description{
Supported States
}
\keyword{datasets}
