% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphiepidata-package.R
\docType{package}
\name{delphiepidata}
\alias{delphiepidata}
\alias{delphiepidata-package}
\title{Query the 'CMU' 'DELPHI' Epidemiological Data 'API'}
\description{
The 'CMU' 'DELPHI' (\url{https://delphi.midas.cs.cmu.edu/}) service
provides an aggregated, central point of access to influenza-like
illness ('ILI') related data sources. Methods are provided to query
all supported endpoints.
}
\details{
\itemize{
\item DELPHI API docs: \url{https://github.com/cmu-delphi/delphi-epidata}
\item DELPHI: \url{https://delphi.midas.cs.cmu.edu/}
\item URL: \url{https://gitlab.com/hrbrmstr/delphiepidata}
\item BugReports: \url{https://gitlab.com/hrbrmstr/delphiepidata/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
