% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{dengue_nowcast}
\alias{dengue_nowcast}
\title{Fetch Delphi's PAHO Dengue nowcast}
\usage{
dengue_nowcast(locations, epiweeks)
}
\arguments{
\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{dephi_states}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}
}
\description{
Fetch Delphi's PAHO Dengue nowcast
}
