% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi-cities.R
\docType{data}
\name{deplhi_cities}
\alias{deplhi_cities}
\title{Supported Cities}
\format{An object of class \code{character} of length 97.}
\usage{
deplhi_cities
}
\description{
Supported Cities
}
\keyword{datasets}
