% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{flusurv}
\alias{flusurv}
\title{Fetch FluSurv data}
\usage{
flusurv(locations, epiweeks, issues, lag)
}
\arguments{
\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{dephi_states}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{issues}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{lag}{a number of weeks}
}
\description{
Fetch FluSurv data
}
