% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{fluview}
\alias{fluview}
\title{Fetch FluView data}
\usage{
fluview(regions, epiweeks, issues, lag, auth)
}
\arguments{
\item{regions}{a \code{list} of \link{delphi_regions}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{issues}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{lag}{a number of weeks}

\item{auth}{password for private imputed data}
}
\description{
Fetch FluView data
}
