% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{fluview_clinical}
\alias{fluview_clinical}
\title{Fetch FluView virological data}
\usage{
fluview_clinical(regions, epiweeks, issues, lag)
}
\arguments{
\item{regions}{a \code{list} of \link{delphi_regions}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{issues}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{lag}{a number of weeks}
}
\description{
Fetch FluView virological data
}
