% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{gft}
\alias{gft}
\title{Fetch Google Flu Trends data}
\usage{
gft(locations, epiweeks)
}
\arguments{
\item{locations}{a \code{list} of \link{delphi_regions}, \link{dephi_states}, and/or \link{delphi_cities}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}
}
\description{
Fetch Google Flu Trends data
}
