% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{ght}
\alias{ght}
\title{Fetch Google Health Trends data}
\usage{
ght(auth, locations, epiweeks, query)
}
\arguments{
\item{auth}{authentication token}

\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{delphi_cities}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{query}{google query}
}
\description{
Fetch Google Health Trends data
}
