% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{nidss_dengue}
\alias{nidss_dengue}
\title{Fetch NIDSS dengue data}
\usage{
nidss_dengue(locations, epiweeks)
}
\arguments{
\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{dephi_states}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}
}
\description{
Fetch NIDSS dengue data
}
