% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{nidss_flu}
\alias{nidss_flu}
\title{Fetch NIDSS flu data}
\usage{
nidss_flu(regions, epiweeks, issues, lag)
}
\arguments{
\item{regions}{a \code{list} of \link{delphi_regions}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{issues}{a \code{list} of issues}

\item{lag}{a number of weeks}
}
\description{
Fetch NIDSS flu data
}
