% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{norostat}
\alias{norostat}
\title{Fetch NoroSTAT data (point data, no min/max)}
\usage{
norostat(auth, location, epiweeks)
}
\arguments{
\item{auth}{auth token}

\item{location}{a single \link{delphi_regions} and/or \link{dephi_states}}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}
}
\description{
Fetch NoroSTAT data (point data, no min/max)
}
