% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{twitter}
\alias{twitter}
\title{Fetch HealthTweets data}
\usage{
twitter(auth, locations, dates, epiweeks)
}
\arguments{
\item{auth}{twitter auth token}

\item{locations}{a \code{list} of \link{delphi_regions} and/or \link{delphi_cities}}

\item{dates}{a \code{list} of dates (\code{YYYYMMDD} format)}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}
}
\description{
Fetch HealthTweets data
}
