% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{wiki}
\alias{wiki}
\title{Fetch Wikipedia access data}
\usage{
wiki(articles, dates, epiweeks, hours)
}
\arguments{
\item{articles}{a \code{list} of \link{delphi_articles}}

\item{dates}{a \code{list} of dates (\code{YYYYMMDD} format)}

\item{epiweeks}{a \code{list} of epiweeks (format for an epiweek is \code{YYYYWW})}

\item{hours}{a \code{list} of hours}
}
\description{
Fetch Wikipedia access data
}
