% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphiepidata-package.R
\docType{package}
\name{delphiepidata}
\alias{delphiepidata}
\alias{delphiepidata-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/delphiepidata}
\item BugReports: \url{https://gitlab.com/hrbrmstr/delphiepidata/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
